Configure frontend single-page-app to work with backend routing.
Added minimal code and configuration to demonstrate how the frontend SPA will work with the backend routing.
All requests to the backend (except those to the path '/api') are redirected to the homepage - which is the root of the SPA.
The SPA handles all other in-app routing using the vue-router, behind the scenes, it makes calls to the backend through the '/api' endpoint.