Add major components of SPA frame.
Cleaned up the API for the AppHeader component.
Added the SideNav component with an API for easy creation of main-section links.
Added SPA routing for the major components like MyFeed and MySchedule, other components can be easily plugged in later.

Todo:
Easy universal themeing using a global SASS file.
Consider using separate component for the user-account-preview on the AppHeader component.