Resolved issue of database .db files not being found in production mode.
Play was not finding the databases when running the server in production mode after `sbt dist` or `sbt stage`. The reason was that play did not package the .db files by default, so the urls we specified in application.conf was wrong at runtime. The solution was to move the .db files into the /conf directory - Play will package any files found here - then we change the urls appropriately in application.conf.

Note however, that this feels very much like a hack to me, the database should ideally be on a separate filesystem location so it is not dependent on the Play project structure.